+++ packages/extension/src/popup.tsx
@@ -1,6 +1,6 @@
 import { createRoot } from 'react-dom/client';
 
-import { App } from './panel/App.jsx';
+import { App } from './ui/App.jsx';
 
 // @ts-expect-error - our root element is not null
 const root = createRoot(document.getElementById('root'));
