+++ packages/extension/src/ui/App.module.css
