+++ packages/extension/src/ui/components/KernelControls.tsx
