+++ packages/extension/src/ui/components/MessagePanel.tsx
