+++ packages/extension/src/ui/hooks/useStatusPolling.ts
