+++ packages/extension/src/ui/services/stream.ts
